minor tweaks

+++ DWT/Final/serial/DWT.h
@@ -8,8 +8,8 @@
 class DWT {
 public:
     // Constructor to initialize the DWT with low-pass and high-pass filters
-    DWT(const float* lpf, const float* hpf, int filter_size);
-
+    DWT(const float* lpf, const float* hpf, size_t filter_size);
+   
     // Function to perform 3D Discrete Wavelet Transform on the input data
     Array3D<float> dwt_3d(const Array3D<float>& data, int levels) const;
